move getDevToolsData
