+++ src/tools/ToolDefinition.ts
@@ -204,7 +204,6 @@ export type Context = Readonly<{
     timeout?: number,
     page?: Page,
   ): Promise<Element>;
-  getDevToolsData(page: ContextPage): Promise<DevToolsData>;
   /**
    * Returns a reqid for a cdpRequestId.
    */
@@ -253,8 +252,8 @@ export type ContextPage = Readonly<{
     toolName: string,
     params: Record<string, unknown>,
     response: Response,
-    context: Context,
   ): Promise<void>;
+  getDevToolsData(): Promise<DevToolsData>;
 }>;
 
 export function defineTool<Schema extends zod.ZodRawShape>(
